Add bloggers controller. Add bloggers logic.

+++ src/features/blogs/application/usecases/delete-blog.usecase.ts
@@ -1,10 +1,14 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
-import { NotFoundException } from '@nestjs/common';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
 import { BlogsRepository } from '../../infrastructure/blogs.repository';
 import { BlogsQueryRepository } from '../../infrastructure/blogs.query-repository';
+import { AccessTokenPayloadDTO } from '../../../auth/api/dto/input/access-token-params.dto';
 
 export class DeleteBlogByIdCommand {
-  constructor(public readonly blogId: number) {}
+  constructor(
+    public readonly blogId: number,
+    public readonly user: AccessTokenPayloadDTO,
+  ) {}
 }
 
 @CommandHandler(DeleteBlogByIdCommand)
@@ -17,11 +21,16 @@ export class DeleteBlogByIdUseCase
   ) {}
 
   async execute(command: DeleteBlogByIdCommand) {
-    const { blogId } = command;
+    const { blogId, user } = command;
     const blog = await this.blogsQueryRepository.findById(blogId);
     if (!blog) {
       throw new NotFoundException(`Blog not found`);
     }
+
+    if (+user.userId !== blog.owner.id) {
+      throw new ForbiddenException(`User ID does not exist`);
+    }
+
     await this.blogsRepository.deleteBlogById(blogId);
   }
 }
